requirements.txt
discord.py[voice]
gTTS
python-dotenv